ItemView for HiGlassViewConfig Item -- Basic
