+++ src/encoded/static/components/item-pages/components/HiGlass/index.js
@@ -2,5 +2,5 @@
 
 export { HiGlassContainer, HiGlassConfigurator } from './HiGlassContainer';
 export { HiGlassPlainContainer } from './HiGlassPlainContainer';
-export { HiGlassFileTabView, HiGlassViewConfigTabView } from './HiGlassTabView';
+export { HiGlassFileTabView } from './HiGlassFileTabView';
 export { HiGlassLocalStorage } from './HiGlassLocalStorage';
